+++ lib/services/provider/font_size_provider.dart
@@ -7,12 +7,12 @@ class FontSizeProvider extends ChangeNotifier {
     _preferences = MyFontSizePreferences();
     getPreferences();
   }
-  late int _fontSize;
+  late double _fontSize;
   late MyFontSizePreferences _preferences;
-  int get fontSize => _fontSize;
+  double get fontSize => _fontSize;
 
   // setting font size value
-  set fontSize(int value) {
+  set fontSize(double value) {
     _fontSize = value;
     _preferences.setFontSize(value);
     notifyListeners();
